Se instalaron los paquetes solicitados por el IDE. Al igual que  se actualizaron los paquetes existentes
Se realizo la instalación de la paquetería Xamarin.AndroidX.MediaRouter y Xamarin.AndroidX.Palette.

Se añadió una línea de código en AssemblyInfo para permitir la conexión HTTP de respuesta de la API.

De igual manera y por motivos de la compatibilidad con la versión 2017 de Visual Studio Community. Se realizo la instalación de Xamarin.Essentials para la compatibilidad de las nuevas aplicaciones con Visual Studio 2019 Community.

+++ WeatherApp/WeatherApp/Models/weatherdata.cs
@@ -37,14 +37,14 @@ namespace WeatherApp.Models
     public class Current
     {
         public string observation_time { get; set; }
-        public float temperature { get; set; }
+        public int temperature { get; set; }
         public int weather_code { get; set; }
         public string[] weather_icons { get; set; }
         public string[] weather_descriptions { get; set; }
         public int wind_speed { get; set; }
         public int wind_degree { get; set; }
         public string wind_dir { get; set; }
-        public float pressure { get; set; }
+        public int pressure { get; set; }
         public float precip { get; set; }
         public int humidity { get; set; }
         public int cloudcover { get; set; }
@@ -53,5 +53,4 @@ namespace WeatherApp.Models
         public int visibility { get; set; }
         public string is_day { get; set; }
     }
-
 }
